intelligent_data_assistant/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
nltk
scipy
scikit-learn
reportlab
python-dotenv
requests
SpeechRecognition
